% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{allow_overwrite}
\alias{allow_overwrite}
\title{Internal Function for Confirming Overwrite}
\usage{
allow_overwrite(path)
}
\arguments{
\item{path}{the directory to create}
}
\value{
logical; \code{TRUE} if the overwrite can proceed, \code{FALSE} if not
}
\description{
Internal Function for Confirming Overwrite
}
