% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{checked_dir_create}
\alias{checked_dir_create}
\title{Internal Function for Creating Directories}
\usage{
checked_dir_create(path, ...)
}
\arguments{
\item{path}{the directory to create}
}
\description{
Internal Function for Creating Directories
}
