% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{data_grid_labels}
\alias{data_grid_labels}
\title{Extract Positions to Place Labels on Grid Lines}
\usage{
data_grid_labels(
  dt,
  axiscol,
  scalefun = scales::pretty_breaks(),
  labelfun = as.character,
  by,
  ...,
  exclude.zero = TRUE
)
}
\arguments{
\item{dt}{a `data.table`: generally intended to be the data for the whole
`ggplot` object.}

\item{axiscol}{a bare variable name: the column in the `data.table` which the
scale will be over.}

\item{scalefun}{a function: should accept a vector of the same type as
represented by `axiscol` and return a vector of that type.}

\item{labelfun}{a function: should accept a vector of the same type as
represented by `axiscol` and return a `character` vector (of labels).}

\item{by}{optional expression of the same type as [data.table::data.table]
`by` argument.}

\item{...}{other named arguments: these will be added as columns to the
returned `data.table`}
}
\description{
Intended for use with [ggplot2] `geom`s `data` argument, uses
`data.table`-like syntax to identify label positions + values.
}
