% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynfever_functions.R
\name{dynfever_dStep}
\alias{dynfever_dStep}
\title{Single Dynamical Fever Step}
\usage{
dynfever_dStep(t, y, parms, ...)
}
\arguments{
\item{t}{the time step}

\item{y}{a named integer vector:
`c(N_human, N_dog, I_human, I_dog, C_human, C_dog)`}

\item{parms}{a list: the model parameters; transformed from
[dynfever_parameters()] to `p_avoid = c(human = ..., dog = ...)` in
[dynfever_simulate()]}

\item{...}{ignored}
}
\value{
a list: conforming to [deSolve::ode()] return values, a
`list(c(dN, dI, dC), FOI)`
}
\description{
A function to compute the next state for the Dynamical Fever
Model; function signature to match the style of functions required by
[deSolve::ode()] family.
}
\details{
n.b. arguments are unchecked, as this is intended for internal use
in [dynfever_simulate()] repeatedly. The arguments are checked in that scope.
}
\examples{
y <- dynfever_pop()
ps <- dynfever_params()
dy <- dynfever_dStep(y = y, parms = ps)
dy
y <- y + dy$states
dynfever_dStep(y = y, parms = ps)
# note, uses random number generator, so outcomes may differ:
dynfever_dStep(y = y, parms = ps)

}
\seealso{
Other dynfever: 
\code{\link{dynfever_incidence}()},
\code{\link{dynfever_params}()},
\code{\link{dynfever_plot_heatmap}()},
\code{\link{dynfever_plot_series}()},
\code{\link{dynfever_plot_simple}()},
\code{\link{dynfever_pop}()},
\code{\link{dynfever_sample}()},
\code{\link{dynfever_simulate}()},
\code{\link{dynfever_summarize}()},
\code{\link{dynfever_toLong}()}
}
\concept{dynfever}
