% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynfever_functions.R
\name{dynfever_params}
\alias{dynfever_params}
\title{Setup Dynamical Fever Parameters}
\usage{
dynfever_params(
  R0 = 2,
  pp = 0.002,
  vax_humans = 0,
  vax_dogs = 0,
  N_d = dynfever_pop()["N_d"]
)
}
\arguments{
\item{R0}{a positive number: $R_0$, the reproductive number of Dynamical
Fever (DF) in a naive dog population}

\item{pp}{a probability: the pairwise probability of dog-human effective
contact}

\item{vax_humans}{a probability: the vaccinated fraction of the human
population}

\item{vax_dogs}{a probability: the vaccinated fraction of the dog population}
}
\value{
a list of `numeric`s:
 - `R0`, the supplied $R_0$
 - `pp`, the supplied effective contact probability
 - `vax_dogs` and `vax_humans`, the vaccinated proportions
 - `p_avoid`, the probabilities for humans and dogs of avoiding DF
}
\description{
Setup Dynamical Fever Parameters
}
\examples{
dynfever_params # to see function contents
dfpar <- dynfever_params()
dfpar
dfpar <- dynfever_params(vax_dogs = 0.5, R0 = 3)
dfpar
# note, will complain about nonsensical parameters:
try(dynfever_params(R0 = -1))
try(dynfever_params(pp = 1.5))
try(dynfever_params(vax_humans = -1.5))
try(dynfever_params(vax_dogs = 1.5))

}
\seealso{
[dynfever_solve()]

Other dynfever: 
\code{\link{dynfever_dStep}()},
\code{\link{dynfever_incidence}()},
\code{\link{dynfever_plot_heatmap}()},
\code{\link{dynfever_plot_series}()},
\code{\link{dynfever_plot_simple}()},
\code{\link{dynfever_pop}()},
\code{\link{dynfever_sample}()},
\code{\link{dynfever_simulate}()},
\code{\link{dynfever_summarize}()},
\code{\link{dynfever_toLong}()}
}
\concept{dynfever}
