% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynfever_plotting.R
\name{dynfever_plot_heatmap}
\alias{dynfever_plot_heatmap}
\title{Dynamical Fever: Compare Vaccine Pops}
\usage{
dynfever_plot_heatmap(dt, countbinwidth = 10)
}
\arguments{
\item{dt}{a [data.table::data.table()]; with columns:
- `sample`, the sample id
- `vax_pop`, the vaccinated population, either "h" or "d"
- `pop`, the measured population, either "h" or "d"
- `vax_cov`, the coverage in the vaccinated population
- `count`, the final size of the outbreak in `pop`}

\item{countbinwidth}{positive integer, how wide to make the bins for
counts?}
}
\value{
a [ggplot2::ggplot()] object
}
\description{
Dynamical Fever: Compare Vaccine Pops
}
\seealso{
Other dynfever: 
\code{\link{dynfever_dStep}()},
\code{\link{dynfever_incidence}()},
\code{\link{dynfever_params}()},
\code{\link{dynfever_plot_series}()},
\code{\link{dynfever_plot_simple}()},
\code{\link{dynfever_pop}()},
\code{\link{dynfever_sample}()},
\code{\link{dynfever_simulate}()},
\code{\link{dynfever_summarize}()},
\code{\link{dynfever_toLong}()}
}
\concept{dynfever}
