% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynfever_plotting.R
\name{dynfever_plot_series}
\alias{dynfever_plot_series}
\title{Plot DynFever Outputs as an Annual Series}
\usage{
dynfever_plot_series(
  dynfever_output = dynfever_sample(n = 5),
  startyear = {
     res <- as.POSIXlt(Sys.Date())
     res$year <- res$year - 5
    
    res$mon <- 0
     res$mday <- 1
     res
 }
)
}
\arguments{
\item{dynfever_output}{see return value of [dynfever_sample()]}

\item{startyear}{the "year 0" corresponding to `sample == 1` outcomes}
}
\description{
Plots a multi-year Dynamical Fever series
}
\seealso{
Other dynfever: 
\code{\link{dynfever_dStep}()},
\code{\link{dynfever_incidence}()},
\code{\link{dynfever_params}()},
\code{\link{dynfever_plot_heatmap}()},
\code{\link{dynfever_plot_simple}()},
\code{\link{dynfever_pop}()},
\code{\link{dynfever_sample}()},
\code{\link{dynfever_simulate}()},
\code{\link{dynfever_summarize}()},
\code{\link{dynfever_toLong}()}
}
\concept{dynfever}
