% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynfever_plotting.R
\name{dynfever_plot_simple}
\alias{dynfever_plot_simple}
\alias{dynfever_plot_empty}
\title{Dynamical Fever Dog & Human Incidence Plot}
\usage{
dynfever_plot_simple(dynfever_output = dynfever_sample(), sample_id)

dynfever_plot_empty()
}
\arguments{
\item{dynfever_output}{the output of [dynfever_simulate()] and [dynfever_sample()];
either the full list, with elements `states` and `params`, both
`data.table`s; or directly one of those two elements. If the input is already
in long format, it will be returned unchanged. See [dynfever_simulate()] and
[dynfever_sample()] for the base output formats}

\item{sample_id}{an integer: a particular sample from outputs to consider}
}
\description{
Dynamical Fever Dog & Human Incidence Plot
}
\seealso{
Other dynfever: 
\code{\link{dynfever_dStep}()},
\code{\link{dynfever_incidence}()},
\code{\link{dynfever_params}()},
\code{\link{dynfever_plot_heatmap}()},
\code{\link{dynfever_plot_series}()},
\code{\link{dynfever_pop}()},
\code{\link{dynfever_sample}()},
\code{\link{dynfever_simulate}()},
\code{\link{dynfever_summarize}()},
\code{\link{dynfever_toLong}()}

Other dynfever: 
\code{\link{dynfever_dStep}()},
\code{\link{dynfever_incidence}()},
\code{\link{dynfever_params}()},
\code{\link{dynfever_plot_heatmap}()},
\code{\link{dynfever_plot_series}()},
\code{\link{dynfever_pop}()},
\code{\link{dynfever_sample}()},
\code{\link{dynfever_simulate}()},
\code{\link{dynfever_summarize}()},
\code{\link{dynfever_toLong}()}
}
\concept{dynfever}
