% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynfever_functions.R
\name{dynfever_pop}
\alias{dynfever_pop}
\title{Initialize a Dynamical Fever Population}
\usage{
dynfever_pop(dogs = 1000L, humans = 1000L)
}
\arguments{
\item{dogs}{an integer: how many dogs in the population?}

\item{humans}{an integer: how many humans in the population?}
}
\value{
a named, non-negative integer vector, length 6: `N_X`, `I_X`, `C_X`
for `X` in `h` (humans) and `d` (dogs), corresponding to the number of
susceptible (`N`), currently infected (`I`), and ever infected (`C`)
individuals in each population.

There is always a single infected dog and no infected humans.
}
\description{
Initialize a Dynamical Fever Population
}
\examples{
dynfever_pop # to see function contents
dfpop <- dynfever_pop()
dfpop
dfpop <- dynfever_pop(dogs = 1e5)
dfpop
# note that non-sense arguments will produce errors
try(dynfever_pop(dogs = -1))
try(dynfever_pop(humans = 1.5))

}
\seealso{
Other dynfever: 
\code{\link{dynfever_dStep}()},
\code{\link{dynfever_incidence}()},
\code{\link{dynfever_params}()},
\code{\link{dynfever_plot_heatmap}()},
\code{\link{dynfever_plot_series}()},
\code{\link{dynfever_plot_simple}()},
\code{\link{dynfever_sample}()},
\code{\link{dynfever_simulate}()},
\code{\link{dynfever_summarize}()},
\code{\link{dynfever_toLong}()}
}
\concept{dynfever}
