% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynfever_functions.R
\name{dynfever_sample}
\alias{dynfever_sample}
\title{Dynamical Fever Sampler}
\usage{
dynfever_sample(
  n = 1L,
  y = dynfever_pop(),
  times = 1L:MAXTIME,
  parms = dynfever_params(N_d = y["N_d"]),
  MAXTIME = 30,
  ref.seed,
  ...
)
}
\arguments{
\item{n}{a positive integer: the number of samples}

\item{y}{a named vector of integers: the initial population of humans & dogs,
both infectious and susceptible}

\item{times}{a series of integers, starting with 1L: the times to report;
result with also include time 0L.}

\item{parms}{a list: either same as return of [dynfever_params()] or a list
of elements that are each the same as return of [dynfever_params()]}

\item{ref.seed}{see [set.seed()]; the initial seed. each sample will be a
specific shift from this seed}
}
\value{
a  pair of `data.table`s, same as [dynfever_simulate()] with added
a `sample` columns.
}
\description{
Dynamical Fever Sampler
}
\examples{
one_run <- dynfever_simulate(seed = 1L)
ten_run <- dynfever_sample(n = 10, ref.seed = 1L)
# note:
all(one_run$states == ten_run$states[sample == 1][, -1])
# dynfever_sample() can also be used with a series of paramters
pars <- seq(.1, .5, by = .1) |> lapply(
 \(vax_d) dynfever_params(vax_dogs = vax_d)
)

five_run <- dynfever_sample(n=5, parms = pars)

}
\seealso{
[dynfever_simulate()]

Other dynfever: 
\code{\link{dynfever_dStep}()},
\code{\link{dynfever_incidence}()},
\code{\link{dynfever_params}()},
\code{\link{dynfever_plot_heatmap}()},
\code{\link{dynfever_plot_series}()},
\code{\link{dynfever_plot_simple}()},
\code{\link{dynfever_pop}()},
\code{\link{dynfever_simulate}()},
\code{\link{dynfever_summarize}()},
\code{\link{dynfever_toLong}()}
}
\concept{dynfever}
