% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynfever_functions.R
\name{dynfever_summarize}
\alias{dynfever_summarize}
\title{Summarize Dynamical Fever Results}
\usage{
dynfever_summarize(dynfever_output)
}
\arguments{
\item{dynfever_output}{the output of [dynfever_simulate()] and [dynfever_sample()];
either the full list, with elements `states` and `params`, both
`data.table`s; or directly one of those two elements. If the input is already
in long format, it will be returned unchanged. See [dynfever_simulate()] and
[dynfever_sample()] for the base output formats}
}
\description{
Reports the peak value and timing for infections, as well as the
final size and outbreak duration, stratified by population.
}
