% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hivTutorial.R
\name{hiv_SI}
\alias{hiv_SI}
\alias{hiv_SIhet}
\alias{hiv_SI4}
\alias{hiv_SI4control}
\alias{hiv_SI4react}
\title{HIV in Harare Models}
\usage{
hiv_SI(t, y, parms)

hiv_SIhet(t, y, parms)

hiv_SI4(t, y, parms)

hiv_SI4control(t, y, parms)

hiv_SI4react(t, y, parms)
}
\description{
These are the five models considered in the \code{\link[=hivtutorial]{hivtutorial()}} interactive
shiny app. They all conform to the requirements for \code{\link[deSolve:ode]{deSolve::ode()}}
function arguments.
}
\details{
\itemize{
\item \code{hiv_SI}: a susceptible-infectious model, with no other processes
\item \code{hiv_SIhet}: \code{hiv_SI}, but with less transmission as more of the
population is infectious, representing higher-risk people being infected
first
\item \code{hiv_SI4}: \code{hiv_SIhet}, but with the infectious compartment represented
a boxcar
\item \code{hiv_SI4control}: \code{hiv_SI4}, but introducing a control program
\item \code{hiv_SI4react}: \code{hiv_SI4}, but with individuals reacting to AIDS deaths
}
}
