% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneityTutorial.R
\name{het.population}
\alias{het.population}
\title{Sample Population with Heterogeneous Contact Rates}
\usage{
het.population(n, beta.mean, beta.var, seed)
}
\arguments{
\item{n}{Number of individuals in the population}

\item{beta.mean}{Mean contact rate}

\item{beta.var}{Variance of contact rate}

\item{seed}{Seed for random number generation}
}
\value{
A numeric vector of contact rates, sorted from highest to lowest
}
\description{
Create a sample of individuals with heterogeneous contact rates
}
\details{
Per `?rgamma`, the gamma distribution is parameterized by `shape`
and `scale` parameters.  These parameters are related to the mean by
\deqn{E(X)=shape*scale} and the variance by \deqn{Var(X)=shape*scale^2}.
}
\examples{
require(ICI3D)
egpop <- het.population(1000, 2, .5)
het.hist(egpop, 2)
}
\seealso{
Other heterogeneity: 
\code{\link{het.hist}()},
\code{\link{het.run}()},
\code{\link{sample_pair}()},
\code{\link{sample_until}()}
}
\concept{heterogeneity}
