% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/het_functions.R
\name{het_finalsize_hist}
\alias{het_finalsize_hist}
\title{Heterogeneous Final Size Histogram}
\usage{
het_finalsize_hist(ts, binwidth = 5, ...)
}
\arguments{
\item{ts}{a [data.table::data.table()]; the return from [het_sample()]}
}
\value{
a [ggplot2::ggplot()] object
}
\description{
Plots a histogram of final sizes from many [het_sample()]s
}
\examples{
hp <- het_population(n = 500, beta_mean = 2, beta_var = 1)
samples <- het_sample(100, hp)
het_finalsize_hist(samples, 10)

}
\seealso{
Other heterogeneity: 
\code{\link{het_contact_hist}()},
\code{\link{het_plot}()},
\code{\link{het_population}()},
\code{\link{het_sample}()},
\code{\link{het_simulate}()}
}
\concept{heterogeneity}
