% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/het_functions.R
\name{het_plot}
\alias{het_plot}
\title{Time Series Plot for Heterogeneity Tutorial}
\usage{
het_plot(ts)
}
\arguments{
\item{ts}{a [data.table()], as yielded by [het_sample()]}
}
\value{
a [ggplot2::ggplot()] object
}
\description{
Produces a standard compartment time series plot for outputs from
[het_sample()].
}
\examples{
het_plot(het_sample(30))

}
\seealso{
Other heterogeneity: 
\code{\link{het_contact_hist}()},
\code{\link{het_finalsize_hist}()},
\code{\link{het_population}()},
\code{\link{het_sample}()},
\code{\link{het_simulate}()}
}
\concept{heterogeneity}
