% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/het_functions.R
\name{het_population}
\alias{het_population}
\title{Heterogeneous Population Generation}
\usage{
het_population(n, beta_mean, beta_var, seed)
}
\arguments{
\item{n}{an integer: the number of individuals to generate}

\item{beta_mean}{a numeric: the mean contact rate}

\item{beta_var}{a numeric: the variance of the contact rate}

\item{seed}{an integer, optional: the random seed to use (if supplied)}
}
\value{
a numeric vector of length `n` of contact rates, sorted in descending
order
}
\description{
Generates a heterogeneous population of contact rates
}
\details{
This function generates a heterogeneous population of contact rates
using a gamma distribution.  The mean and variance of the distribution are
transformed into the shape and scale parameters for [rgamma()].
}
\examples{
hp <- het_population(n = 1000, beta_mean = 1, beta_var = 1)
hist(hp)

}
\seealso{
Other heterogeneity: 
\code{\link{het_plot_hist}()},
\code{\link{sample_pair}()},
\code{\link{sample_until}()}
}
\concept{heterogeneity}
