% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/het_functions.R
\name{het_sample}
\alias{het_sample}
\title{Sample [het_simulate()]}
\usage{
het_sample(
  n,
  mxdst = het_population(n = 300, beta_mean = 1, beta_var = 0.5),
  params = list(tmax = 5, recovery_rate = 1, waning_rate = 0)
)
}
\arguments{
\item{n}{a positive integer, the number of samples to simulate}

\item{mxdst}{individual contact rates, as returned by [het_population()]}

\item{params}{a list, with `tmax`, `recovery_rate`, and `waning_rate` keys}
}
\description{
Run [het_simulate()] many times with the same parameter combinations.
}
\examples{
hetruns <- het_sample(10)
hetruns[day == 5.0]

}
\seealso{
Other heterogeneity: 
\code{\link{het_contact_hist}()},
\code{\link{het_finalsize_hist}()},
\code{\link{het_population}()},
\code{\link{het_simulate}()},
\code{\link{sample_pair}()},
\code{\link{sample_until}()}
}
\concept{heterogeneity}
