% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/het_functions.R
\name{het_simulate}
\alias{het_simulate}
\title{Simulate Heterogeneous Transmission Model}
\usage{
het_simulate(mxdst, params, seed)
}
\arguments{
\item{mxdst}{individual contact rates, as returned by [het_population()]}

\item{params}{a list, with `tmax`, `recovery_rate`, and `waning_rate` keys}
}
\description{
Run the heterogeneous transmission model for a particular parameter
combination.
}
\seealso{
Other heterogeneity: 
\code{\link{het_contact_hist}()},
\code{\link{het_finalsize_hist}()},
\code{\link{het_population}()},
\code{\link{het_sample}()},
\code{\link{sample_pair}()},
\code{\link{sample_until}()}
}
\concept{heterogeneity}
