% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICI3D-package.R
\docType{package}
\name{ICI3D-package}
\alias{ICI3D}
\alias{ICI3D-package}
\title{ICI3D: Materials for the International Clinics on Infectious Disease Dynamics and Data (ICI3D) Workshops}
\description{
The ICI3D package provides course material--data and rshiny demonstration applications-- for the Clinic on Dynamical Approaches to Infectious Disease Data (DAIDD) and the Clinic on the Meaningful Modeling of Epidemiological Data (MMED), run by the [ICI3D Program](https://www.ici3d.org).
}
\author{
\strong{Maintainer}: Carl Pearson \email{carl.ab.pearson@gmail.com}

Other contributors:
\itemize{
  \item International Clinics on Infectious Disease Dynamics and Data \email{admin@ici3d.org} [copyright holder]
}

}
\keyword{internal}
