% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{launch}
\alias{launch}
\title{Launch an ICI3D Tutorial}
\usage{
launch(tutorial = c("dynfever", "heterogeneity"))
}
\arguments{
\item{tutorial}{a string: the tutorial name (must be among the defaults)}
}
\description{
Launch an ICI3D Tutorial
}
