% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/het_functions.R
\name{sample_until}
\alias{sample_until}
\title{Sample Event Times Until a Maximum Time}
\usage{
sample_until(maxval, meantime, sampler, ..., oversample = 0.1, verbose = FALSE)
}
\arguments{
\item{maxval}{a positive numeric: the maximum time to simulate until}

\item{meantime}{a positive numeric: the mean time between events}

\item{sampler}{a function: the sampling distribution for time between events}

\item{...}{any parameter arguments to `sampler`}

\item{oversample}{a numeric: the amount to oversample by when generating}

\item{verbose}{Print progress messages?}
}
\value{
A numeric vector of event times: the cumulative sum of draws from
`sampler` until the last value exceeds `maxval`. That sum is then trimmed to
values `<= maxval`. The first value in the vector is always 0.
}
\description{
Sample Event Times Until a Maximum Time
}
\examples{
draws <- sample_until(n = 100, meantime = 1, rexp, rate = 1)
cumsum(draws)

}
\seealso{
Other heterogeneity: 
\code{\link{het_contact_hist}()},
\code{\link{het_finalsize_hist}()},
\code{\link{het_population}()},
\code{\link{het_sample}()},
\code{\link{het_simulate}()},
\code{\link{sample_pair}()}
}
\concept{heterogeneity}
