% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{scripts}
\alias{scripts}
\title{Create a local copy of the tutorials.}
\usage{
scripts(
  path = file.path(if (.Platform$OS.type == "windows") {
     Sys.getenv("USERPROFILE")

    } else {
     "~"
 }, "Downloads", "ICI3D"),
  overwrite = FALSE,
  what = c("scripts", "solutions")
)
}
\arguments{
\item{path, }{string; the path to enclosing directory. If this directory does
not exist, will create it}

\item{overwrite, }{logical; overwrite existing files (corresponding to the
course scripts)?}

\item{what, }{string: the "scripts" or the "solutions"?}
}
\value{
string or \code{NULL}; non-\code{NULL} indicates successful creation + copy and
is the top level root of the course material.
}
\description{
Create a local copy of the tutorials.
}
\details{
This function creates a local copy of the R scripts associated with
the ICI3D short courses, for students to edit and run during learning
activities. These scripts should only require the \code{ICI3D} package and
associated dependencies.

The created directory contains several subdirectories, each corresponding to
a different session. Within each directory, the exercise files are generally
named \code{NN_short_description.R} (where NN is 00, 01, etc - corresponding to
the expected order of completion, with the 00 session being background /
preparatory content).

If used with default \code{what} argument, you will get a "scripts" subdirectory,
and if you ask for \code{what = "solutions"} there will be a "solutions"
subdirectory. Both will contain sessions by name, scripts by order and
shortname, without and with solutions filled in, respectively.
}
\examples{
tardir <- ICI3D::scripts()
list.files(tardir, recursive = TRUE)

}
