% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{seq_between}
\alias{seq_between}
\title{Sequence from Range}
\usage{
seq_between(r, ...)
}
\arguments{
\item{r}{an output from [range()]}

\item{...}{
  Arguments passed on to \code{\link[base:range]{base::range}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Returns as sequence between two ends as returned by [range()].
}
\seealso{
Other utilities: 
\code{\link{seq_parms}()},
\code{\link{theyear}()}
}
\concept{utilities}
