% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{seq_parms}
\alias{seq_parms}
\title{Manages a Sequence of Parameters}
\usage{
seq_parms(ref.parms, n)
}
\arguments{
\item{n}{a positive integer: the number of parameter combinations to return}

\item{ref.seed}{a named list OR an unnamed list of length equal to `n`:
respectively
 - the parameter values to be replicated into a sequence
 - the sequence of parameter values}
}
\value{
depending on the input format:
 - a list of clones of parameters
 - the same sequence of parameters provided
}
\description{
Returns a series of parameters from a flexible set of
inputs.
}
\seealso{
Other utilities: 
\code{\link{seq_between}()},
\code{\link{theyear}()}
}
\concept{utilities}
