% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{seq_seeds}
\alias{seq_seeds}
\title{Manages a Sequence of Seeds}
\usage{
seq_seeds(ref.seed, n)
}
\arguments{
\item{ref.seed}{a value that will be coerced to integer OR an object with
length equal to `n` OR `NULL`: either, respectively
 - the starting point for the sequence of seeds
 - the sequence of seeds
 - the indicator to *not* set seeds}

\item{n}{a positive integer: the number of seeds to return}
}
\value{
depending on the input format:
 - a sequence of seeds, `ref.seed + 1L:n - 1L`
 - the same sequence of seeds provided
 - a list of length `n`, with all `NULL` elements
}
\description{
Returns a series of random number seeds from a flexible set of
inputs.
}
\examples{
seq_seeds # to view function internals
tenseeds <- seq_seeds(0, 10)
tenseeds
tenmoreseeds <- seq_seeds(0:9, 10)
tenmoreseeds
# note, won't work with nonsense input:
try(seq_seed(0:10, 10))
# intended for use in `apply`-family invocations of functions that optionally
# want a seed:
seq_seed(n=10)

}
