% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{theyear}
\alias{theyear}
\title{Get Year}
\usage{
theyear(date = Sys.Date())
}
\arguments{
\item{date}{a object coerce-able via [as.Date()]}
}
\value{
an integer, the 4 digit year corresponding to `date`
}
\description{
Gets the year from a date (the current date, by default).
}
\seealso{
Other utilities: 
\code{\link{seq_between}()},
\code{\link{seq_parms}()}
}
\concept{utilities}
