% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneityTutorial.R
\name{sample_pair}
\alias{sample_pair}
\title{Sample Pairs of Distinct Indices}
\usage{
sample_pair(n, prob, verbose = FALSE)
}
\arguments{
\item{n}{positive integer: number of pairs to sample}

\item{prob}{numeric vector: the population weights for sampling each index;
if length(prob) == 1, then treated as the population size and all indices
equally weighted}

\item{verbose}{Print progress messages?}
}
\value{
A list with two elements, `source` and `sink`, each of length `n`,
corresponding to the sampled pairs indices
}
\description{
Sample Pairs of Distinct Indices
}
\seealso{
Other heterogeneity: 
\code{\link{het.hist}()},
\code{\link{het.population}()},
\code{\link{het.run}()},
\code{\link{sample_until}()}
}
\concept{heterogeneity}
