% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneityTutorial.R
\name{sample_until}
\alias{sample_until}
\title{Sample Event Times Until a Maximum Time}
\usage{
sample_until(maxval, meantime, sampler, ..., oversample = 0.1, verbose = FALSE)
}
\arguments{
\item{maxval}{a positive numeric: the maximum time to simulate until}

\item{meantime}{a positive numeric: the mean time between events}

\item{sampler}{a function: the sampling distribution for time between events}

\item{...}{any parameter arguments to `sampler`}
}
\value{
A numeric vector of event times: the cumulative sum of draws from
`sampler` until the last value exceeds `maxval`. That sum is then trimmed to
values `<= maxval`. The first value in the vector is always 0.
}
\description{
Sample Event Times Until a Maximum Time
}
\seealso{
Other heterogeneity: 
\code{\link{het.hist}()},
\code{\link{het.population}()},
\code{\link{het.run}()},
\code{\link{sample_pair}()}
}
\concept{heterogeneity}
